Update printout format

+++ example/bfp_demo/main.c
@@ -47,7 +47,7 @@ static void print_vector(
   // First, the raw mantissas
   printf("%s = [", name);
   for(unsigned int k = 0; k < vect->length; k++)
-    printf("%ld, ", vect->data[k]);
+    printf("%d, ", (int)vect->data[k]);
   printf("] * 2**(%d)\n", vect->exp);
 
   // Next, the float equivalent
